+++ StabilityMatrix.Core/Models/Packages/Extensions/ExtensionManifest.cs
@@ -1,3 +1,6 @@
 namespace StabilityMatrix.Core.Models.Packages.Extensions;
 
-public record ExtensionManifest(Uri Uri);
+public record ExtensionManifest(Uri Uri)
+{
+    public static implicit operator ExtensionManifest(string uri) => new(new Uri(uri, UriKind.Absolute));
+}